+++ lib/ui/wellness/WellnessResourcesContentWidget.dart
@@ -17,12 +17,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
-import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/DeepLink.dart';
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/wellness/WellnessHomePanel.dart';
 import 'package:illinois/ui/widgets/FavoriteButton.dart';
-import 'package:illinois/ui/widgets/LinkButton.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/service/assets.dart';
 import 'package:rokwire_plugin/service/localization.dart';
@@ -35,11 +33,7 @@ class WellnessResourcesContentWidget extends StatefulWidget {
   static const String wellnessCategoryKey = 'resources';
   final String wellnessCategory = wellnessCategoryKey;
 
-  final bool showHeader;
-  final bool showOnlyFavorites;
-  final int? limit;
-
-  WellnessResourcesContentWidget({this.showHeader = true, this.showOnlyFavorites = false, this.limit });
+  WellnessResourcesContentWidget();
 
   @override
   State<WellnessResourcesContentWidget> createState() => _WellnessResourcesContentWidgetState();
@@ -49,7 +43,6 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
 
   List<dynamic>? _commands;
   Map<String, dynamic>? _strings;
-  int? _limit;
 
   @override
   void initState() {
@@ -91,23 +84,21 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
   }
 
   Widget _buildContent() {
-    _limit = widget.limit;
-    return Padding(padding: EdgeInsets.symmetric(horizontal: (widget.limit == null) ? 16 : 0), child: Column(children: [
+    return Padding(padding: EdgeInsets.symmetric(horizontal: 16), child: Column(children: [
       _buildHeader(),
       _buildLargeButtonsContainer(),
       _buildRegularButtonsContainer(),
-      _buildViewAllButton(),
     ]));
   }
 
   Widget _buildHeader() {
-    return widget.showHeader ? Padding(padding: EdgeInsets.only(left: 5, bottom: 10, right: 5), child:
+    return Padding(padding: EdgeInsets.only(left: 5, bottom: 10, right: 5), child:
       Row(crossAxisAlignment: CrossAxisAlignment.center, children: [
         Text(Localization().getStringEx('panel.wellness.resources.header.label', 'Wellness Resources'),
           style: TextStyle(color: Styles().colors!.fillColorPrimary, fontSize: 22, fontFamily: Styles().fontFamilies!.extraBold)
         ),
       ]),
-    ) : Container();
+    );
   }
 
   Widget _buildLargeButtonsContainer() {
@@ -120,28 +111,21 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
           if (type == 'large') {
             String? id = JsonUtils.stringValue(command['id']);
             Favorite favorite = WellnessFavorite(id, category: widget.wellnessCategory);
-            if (!widget.showOnlyFavorites || (Auth2().prefs?.isFavorite(favorite) ?? false)) {
-              if ((_limit == null) || (0 < _limit!)) {
-                if (widgetList.isNotEmpty) {
-                  widgetList.add(Container(height: (widget.limit != null) ? 3 : 8,));
-                }
-                widgetList.add(_buildLargeButton(
-                  label: _getString(id),
-                  favorite: favorite,
-                  hasExternalLink: UrlUtils.isWebScheme(JsonUtils.stringValue(command['url'])),
-                  onTap: () => _onCommand(command),
-                ));
-                if (_limit != null) {
-                  _limit = _limit! - 1;
-                }
-              }
+            if (widgetList.isNotEmpty) {
+              widgetList.add(Container(height: 8));
             }
+            widgetList.add(WellnessLargeResourceButton(
+              label: _getString(id),
+              favorite: favorite,
+              hasExternalLink: UrlUtils.isWebScheme(JsonUtils.stringValue(command['url'])),
+              onTap: () => _onCommand(command),
+            ));
           }
         }
       }
     }
 
-    widgetList.add(Container(height: (widget.limit == null) ? 16 : ((0 < (_limit ?? 0) ? 8 : 0)),));
+    widgetList.add(Container(height: 16,));
     
     return Column(children: widgetList);
   }
@@ -159,19 +143,12 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
             }
             String? id = JsonUtils.stringValue(command['id']);
             Favorite favorite = WellnessFavorite(id, category: widget.wellnessCategory);
-            if (!widget.showOnlyFavorites || (Auth2().prefs?.isFavorite(favorite) ?? false)) {
-              if ((_limit == null) || (0 < _limit!)) {
-                widgetList.add(_buildRegularButton(
-                  label: _getString(id),
-                  favorite: favorite,
-                  hasExternalLink: UrlUtils.isWebScheme(JsonUtils.stringValue(command['url'])),
-                  onTap: () => _onCommand(command),
-                ));
-                if (_limit != null) {
-                  _limit = _limit! - 1;
-                }
-              }
-            }
+            widgetList.add(WellnessRegularResourceButton(
+              label: _getString(id),
+              favorite: favorite,
+              hasExternalLink: UrlUtils.isWebScheme(JsonUtils.stringValue(command['url'])),
+              onTap: () => _onCommand(command),
+            ));
           }
         }
       }
@@ -181,53 +158,6 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
       Column(children: widgetList)
     );
 
-  }
-
-  Widget _buildLargeButton({String? label, Favorite? favorite, bool hasExternalLink = false, void Function()? onTap}) {
-    return Container(decoration: BoxDecoration(color: Styles().colors!.white, border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1), borderRadius: BorderRadius.circular(5)), child:
-      InkWell(onTap: onTap, child:
-        Padding(padding: EdgeInsets.only(left: 16), child:
-          Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
-            Expanded(child:
-              Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
-                Text(label ?? '', style: TextStyle(color: Styles().colors!.fillColorPrimary, fontFamily: Styles().fontFamilies!.bold, fontSize: 18)),
-              ),
-            ),
-            hasExternalLink ? Padding(padding: EdgeInsets.only(left: 6, top: 18, bottom: 18), child:
-              Image.asset('images/external-link.png', color: Styles().colors!.mediumGray)
-            ) : Container(),
-            FavoriteButton(favorite: favorite, style: FavoriteIconStyle.Button, padding: EdgeInsets.only(left: 8, right: 16, top: 16, bottom: 16),)
-          ]),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildRegularButton({String? label, Favorite? favorite, bool hasExternalLink = true, void Function()? onTap}) {
-    return InkWell(onTap: onTap, child:
-      Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
-        FavoriteButton(favorite: favorite, style: FavoriteIconStyle.Button, padding: EdgeInsets.only(left: 16, right: 8, top: 16, bottom: 16)),
-        Expanded(child:
-          Padding(padding: EdgeInsets.symmetric(vertical: 17), child:
-            Text(label ?? '', style: TextStyle(color: Styles().colors!.fillColorPrimary, fontFamily: Styles().fontFamilies!.bold, fontSize: 16))
-          ),
-        ),
-        hasExternalLink ? Padding(padding: EdgeInsets.only(left: 8, top: 18, bottom: 18), child:
-          Image.asset('images/external-link.png', color: Styles().colors!.mediumGray)
-        ) : Container(),
-        Padding(padding: EdgeInsets.only(left: 8, right: 16, top: 18, bottom: 18), child:
-          Image.asset('images/chevron-right.png')
-        ),
-      ]),
-    );
-  }
-
-  Widget _buildViewAllButton() {
-    return (_limit == 0) ? LinkButton(
-        title: Localization().getStringEx('panel.wellness.resources.button.all.title', 'View All'),
-        hint: Localization().getStringEx('panel.wellness.resources.button.all.hint', 'Tap to view all groups'),
-        onTap: _onViewAll,
-      ) : Container();
   }
 
   void _initContent() {
@@ -249,11 +179,6 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
     return null;
   }
 
-  void _onViewAll() {
-    Analytics().logSelect(target: "HomeWellnessResourcesWidget View All");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(content: WellnessContent.resources,)));
-  }
-
   void _onCommand(Map<String, dynamic> command) {
     Analytics().logSelect(target: _getString(JsonUtils.stringValue(command['id']), languageCode: Localization().defaultLocale?.languageCode),);
     _launchUrl(JsonUtils.stringValue(command['url']));
@@ -271,5 +196,71 @@ class _WellnessResourcesContentWidgetState extends State<WellnessResourcesConten
         launch(url!);
       }
     }
+  }
+}
+
+class WellnessLargeResourceButton extends StatelessWidget {
+  final String? label;
+  final Favorite? favorite;
+  final bool hasExternalLink;
+  final void Function()? onTap;
+
+  WellnessLargeResourceButton({Key? key, this.label, this.favorite, this.hasExternalLink = false, this.onTap}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(decoration: BoxDecoration(color: Styles().colors!.white, border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1), borderRadius: BorderRadius.circular(5)), child:
+      InkWell(onTap: onTap, child:
+        Padding(padding: EdgeInsets.only(left: 16), child:
+          Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+            Expanded(child:
+              Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
+                Text(label ?? '', style: TextStyle(color: Styles().colors!.fillColorPrimary, fontFamily: Styles().fontFamilies!.bold, fontSize: 18)),
+              ),
+            ),
+            hasExternalLink ? Padding(padding: EdgeInsets.only(left: 6, top: 18, bottom: 18), child:
+              Image.asset('images/external-link.png', color: Styles().colors!.mediumGray)
+            ) : Container(),
+            FavoriteButton(favorite: favorite, style: FavoriteIconStyle.Button, padding: EdgeInsets.only(left: 8, right: 16, top: 16, bottom: 16),)
+          ]),
+        ),
+      ),
+    );
+  }
+}
+
+class WellnessRegularResourceButton extends StatelessWidget {
+  final String? label;
+  final Favorite? favorite;
+  final bool hasExternalLink;
+  final bool hasBorder;
+  final void Function()? onTap;
+
+  WellnessRegularResourceButton({Key? key, this.label, this.favorite, this.hasExternalLink = false, this.hasBorder = false, this.onTap}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return hasBorder ? Container(decoration: BoxDecoration(color: Styles().colors!.white, border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1), borderRadius: BorderRadius.circular(5)), child:
+      _buildInterior()
+    ) : _buildInterior();
+  }
+
+  Widget _buildInterior() {
+    return InkWell(onTap: onTap, child:
+      Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+        FavoriteButton(favorite: favorite, style: FavoriteIconStyle.Button, padding: EdgeInsets.only(left: 16, right: 8, top: 16, bottom: 16)),
+        Expanded(child:
+          Padding(padding: EdgeInsets.symmetric(vertical: 17), child:
+            Text(label ?? '', style: TextStyle(color: Styles().colors!.fillColorPrimary, fontFamily: Styles().fontFamilies!.bold, fontSize: 16))
+          ),
+        ),
+        hasExternalLink ? Padding(padding: EdgeInsets.only(left: 8, top: 18, bottom: 18), child:
+          Image.asset('images/external-link.png', color: Styles().colors!.mediumGray)
+        ) : Container(),
+        Padding(padding: EdgeInsets.only(left: 8, right: 16, top: 18, bottom: 18), child:
+          Image.asset('images/chevron-right.png')
+        ),
+      ]),
+    );
   }
 }